refactor: Refactor Atomic Design
