+++ src/app/(admin)/_components/statusBanner/OverviewStatBox.tsx
@@ -2,8 +2,8 @@
 
 import React, { useState } from "react";
 import { useRouter } from "next/navigation";
-import StatusBanner from "./StatusBanner";
 import StatusOptions from "./StatusOptions";
+import StatusBanner from "./StatusBanner";
 
 interface OverviewStatBoxProps {
   title: string;
@@ -27,68 +27,7 @@ const OverviewStatBox = ({ title }: OverviewStatBoxProps) => {
         </h1>
         <StatusOptions selected={selected} handleOption={handleOption} />
       </div>
-      <div className="w-full h-[142px] bg-gray1 rounded-[10px] flex items-center justify-center gap-4">
-        <StatusBanner
-          meta={{
-            title: "가입자수",
-            count: 8,
-            totalCount: 165,
-            percentage: 8.5,
-          }}
-          icon={{
-            mainIcon: "USER_JOIN_COUNT",
-            statusIcon: false,
-          }}
-        />
-        <StatusBanner
-          meta={{
-            title: "탈퇴자수",
-            count: 8,
-            totalCount: 165,
-            percentage: 15,
-          }}
-          icon={{
-            mainIcon: "USER_JOIN_COUNT",
-            statusIcon: false,
-          }}
-        />
-        <StatusBanner
-          meta={{
-            title: "탈퇴자수",
-            count: 8,
-            totalCount: 165,
-            percentage: 15,
-          }}
-          icon={{
-            mainIcon: "USER_JOIN_COUNT",
-            statusIcon: false,
-          }}
-        />
-        <StatusBanner
-          meta={{
-            title: "경고",
-            count: 8,
-            totalCount: 165,
-            percentage: 12.5,
-          }}
-          icon={{
-            mainIcon: "USER_JOIN_COUNT",
-            statusIcon: true,
-          }}
-        />
-        <StatusBanner
-          meta={{
-            title: "정지",
-            count: 8,
-            totalCount: 65,
-            percentage: 12.5,
-          }}
-          icon={{
-            mainIcon: "USER_JOIN_COUNT",
-            statusIcon: false,
-          }}
-        />
-      </div>
+      <StatusBanner />
     </div>
   );
 };
